refactor: 유즈케이스를 도메인 레이어로 이동
