+++ android/app/src/main/java/com/zzang/chongdae/di/module/UseCaseDependencyModule.kt
@@ -3,12 +3,12 @@ package com.zzang.chongdae.di.module
 import com.zzang.chongdae.di.annotations.CheckAlreadyLoggedInUseCaseQualifier
 import com.zzang.chongdae.di.annotations.PostLoginUseCaseQualifier
 import com.zzang.chongdae.di.annotations.PostOfferingUseCaseQualifier
-import com.zzang.chongdae.presentation.view.login.usecase.CheckIfAlreadyLoggedInUseCase
-import com.zzang.chongdae.presentation.view.login.usecase.CheckIfAlreadyLoggedInUseCaseImpl
-import com.zzang.chongdae.presentation.view.login.usecase.PostLoginUseCase
-import com.zzang.chongdae.presentation.view.login.usecase.PostLoginUseCaseImpl
-import com.zzang.chongdae.presentation.view.write.usecase.PostOfferingUseCase
-import com.zzang.chongdae.presentation.view.write.usecase.PostOfferingUseCaseImpl
+import com.zzang.chongdae.domain.usecase.login.CheckIfAlreadyLoggedInUseCase
+import com.zzang.chongdae.domain.usecase.login.CheckIfAlreadyLoggedInUseCaseImpl
+import com.zzang.chongdae.domain.usecase.login.PostLoginUseCase
+import com.zzang.chongdae.domain.usecase.login.PostLoginUseCaseImpl
+import com.zzang.chongdae.domain.usecase.write.PostOfferingUseCase
+import com.zzang.chongdae.domain.usecase.write.PostOfferingUseCaseImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
