+++ android/app/src/main/java/com/zzang/chongdae/domain/usecase/login/CheckIfAlreadyLoggedInUseCase.kt
@@ -1,4 +1,4 @@
-package com.zzang.chongdae.presentation.view.login.usecase
+package com.zzang.chongdae.domain.usecase.login
 
 interface CheckIfAlreadyLoggedInUseCase {
     suspend operator fun invoke(): Boolean
