+++ android/app/src/main/java/com/zzang/chongdae/presentation/view/login/LoginViewModel.kt
@@ -3,14 +3,13 @@ package com.zzang.chongdae.presentation.view.login
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.zzang.chongdae.common.datastore.UserPreferencesDataStore
 import com.zzang.chongdae.common.handler.Result
 import com.zzang.chongdae.di.annotations.CheckAlreadyLoggedInUseCaseQualifier
 import com.zzang.chongdae.di.annotations.PostLoginUseCaseQualifier
 import com.zzang.chongdae.presentation.util.MutableSingleLiveData
 import com.zzang.chongdae.presentation.util.SingleLiveData
-import com.zzang.chongdae.presentation.view.login.usecase.CheckIfAlreadyLoggedInUseCase
-import com.zzang.chongdae.presentation.view.login.usecase.PostLoginUseCase
+import com.zzang.chongdae.domain.usecase.login.CheckIfAlreadyLoggedInUseCase
+import com.zzang.chongdae.domain.usecase.login.PostLoginUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
